a11y-keyboard: Manually beep for togglekeys
This allows a customizable and possibly different beep sequence when a
togglekey is enabled or disabled.  This is very useful for the user to
know for sure whether the feature got enabled or disabled.

Back in the days of buzzers, XKB was supposed to use different sounds
for each of those, but this is no longer the case now in the vast
majority of setups the beeps are intercepted and use a single recorded
sound.

XKB beeps are also unfortunately not configurable, although they
possibly should be on paper: in theory, one could alter the bell used
by listening to `XkbBellNotify` events, which provides a way to
discriminate bells through a name.  Unfortunately XKB's togglekeys
seems to suffer from a bug (?) for a long time, in that it will always
ring the `AX_IndicatorOn` bell if there is *at least one* indicator on
at the moment the bell is rung, and `AX_IndicatorOff` if and only if
*all* indicators are off.  This makes it virtually useless as it is not
possible to discriminate between an indicator getting turned on or off
in most cases, especially with NumLock which often stays always on.
Given this behavior dates at the very least as far back as X.org 1.16
which is from 2014, it probably is not very realistic to rely on a fix.

So instead implement togglekeys on our end by listening to the
`XkbIndicatorStateNotify` events.